Move service component to internal.component package.
